backend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
bcrypt==4.0.1
requests==2.31.0
boto3==1.28.55
python-dotenv==1.0.0
Werkzeug==2.3.7
groq>=0.5.0
pydantic>=2.0.0 # Added for recipe service data validation/models

# AWS Lambda integration
aws-wsgi==0.2.7
aws-lambda-powertools==2.26.0

# Serverless plugins (dev dependencies - install via serverless framework, not pip)
# serverless-python-requirements
# serverless-domain-manager 